chore: split locator files into smaller components
